requirements.txt
flask==3.1.0
flask-cors==5.0.1
sentence-transformers==3.4.1
faiss-cpu==1.9.0.post1
numpy==2.2.3
requests==2.32.3
python-dotenv==1.0.1
